# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eBPFsentinel Agent API",
    "description": "REST API for the eBPFsentinel network security agent.\n\nProvides endpoints for managing firewall rules, IPS/IDS policies, rate limiting, L7 filtering, threat intelligence, alerts, and audit logs.\n\nMost endpoints require authentication via JWT Bearer token or API key.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "v0.0.0-dev"
  },
  "paths": {
...
    "/api/v1/ddos/attacks": {
      "get": {
        "tags": [
          "DDoS"
        ],
        "summary": "`GET /api/v1/ddos/attacks` - list active `DDoS` attacks.",
        "operationId": "ddos_attacks",
        "responses": {
          "200": {
            "description": "Active attacks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DdosAttackResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/ddos/attacks/history": {
      "get": {
        "tags": [
          "DDoS"
        ],
        "summary": "`GET /api/v1/ddos/attacks/history` - attack history.",
        "operationId": "ddos_history",
        "responses": {
          "200": {
            "description": "Attack history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DdosAttackResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/ddos/policies": {
      "get": {
        "tags": [
          "DDoS"
        ],
        "summary": "`GET /api/v1/ddos/policies` - list `DDoS` policies.",
        "operationId": "list_ddos_policies",
        "responses": {
          "200": {
            "description": "DDoS policies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DdosPolicyResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "DDoS"
        ],
        "summary": "`POST /api/v1/ddos/policies` - create a `DDoS` policy.",
        "operationId": "create_ddos_policy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDdosPolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Policy created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DdosPolicyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/ddos/policies/{id}": {
      "delete": {
        "tags": [
          "DDoS"
        ],
        "summary": "`DELETE /api/v1/ddos/policies/{id}` - delete a `DDoS` policy.",
        "operationId": "delete_ddos_policy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Policy identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Policy deleted"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Policy not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/ddos/status": {
      "get": {
        "tags": [
          "DDoS"
        ],
        "summary": "`GET /api/v1/ddos/status` - `DDoS` protection status.",
        "operationId": "ddos_status",
        "responses": {
          "200": {
            "description": "DDoS status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DdosStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDdosPolicyRequest": {
        "type": "object",
        "required": [
          "id",
          "attack_type",
          "detection_threshold_pps"
        ],
        "properties": {
          "attack_type": {
            "type": "string"
          },
          "auto_block_duration_secs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "country_thresholds": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "detection_threshold_pps": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "mitigation_action": {
            "type": "string"
          }
        }
      },
...
      "DdosAttackResponse": {
        "type": "object",
        "required": [
          "id",
          "attack_type",
          "status",
          "start_time_ns",
          "peak_pps",
          "current_pps",
          "total_packets",
          "source_count"
        ],
        "properties": {
          "attack_type": {
            "type": "string"
          },
          "current_pps": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string"
          },
          "peak_pps": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "source_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "start_time_ns": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "status": {
            "type": "string"
          },
          "total_packets": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "DdosPolicyResponse": {
        "type": "object",
        "required": [
          "id",
          "attack_type",
          "detection_threshold_pps",
          "mitigation_action",
          "auto_block_duration_secs",
          "enabled"
        ],
        "properties": {
          "attack_type": {
            "type": "string"
          },
          "auto_block_duration_secs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "detection_threshold_pps": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "mitigation_action": {
            "type": "string"
          }
        }
      },
      "DdosStatusResponse": {
        "type": "object",
        "required": [
          "enabled",
          "active_attacks",
          "total_mitigated",
          "policy_count"
        ],
        "properties": {
          "active_attacks": {
            "type": "integer",
            "minimum": 0
          },
          "enabled": {
            "type": "boolean"
          },
          "policy_count": {
            "type": "integer",
            "minimum": 0
          },
          "total_mitigated": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `RULE_NOT_FOUND`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key configured in agent"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained via OIDC or static configuration"
      }
    }
  },
...
}